pyproject.toml
[project]
name = "flask-admin-tabler"
version = "2026.2.4"
description = "Tabler UI theme for Flask-Admin"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "flask-admin>=2.0.2",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.urls]
Source = "https://github.com/mariofix/flask-admin-tabler"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
src = ["flask_admin_tabler"]
fix = true
show-fixes = true

[tool.ruff.lint]
select = [
    "B",   # flake8-bugbear
    "E",   # pycodestyle error
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "W",   # pycodestyle warning
]

[tool.ruff.lint.isort]
force-single-line = true
order-by-type = false
